[+] Добавлена команда directIO для чтения данных QR кода.     Данные возвращаются в параметре pString.     Вызывать нужно после метода endFiscalReceipt.     DIO_FS_READ_QR_CODE_DATA = 80;

+++ Source/SmFiscalPrinter/Units/PrinterTypes.pas
@@ -1167,8 +1167,8 @@ type
 
   TFSCloseReceiptResult2 = record
     Change: Int64;
-    DocNumber: DWORD;
-    MacValue: DWORD;
+    DocNumber: Int64;
+    MacValue: Int64;
     DocDate: TPrinterDate;
     DocTime: TPrinterTime;
   end;
